Some initial ideas
Thought it made sense to use the JsonResponse class instead of
serializing for HttpResponse. Also added some validation on the
request parameters and added a silly handler for adding messages to
responses. Thought the silly handler might be useful later as the
response messages get more complicated.

Also changed 200 to 201 on the create endpoint in the
spirit of rest, but its not super important actually.

Tests are failing, didn't have time to look into it.